json-processor/requirements.txt
google-generativeai   # Gemini SDK (official)
python-dotenv         # Load GOOGLE_API_KEY from .env
tqdm                  # Progress bar (optional, helpful)
